Fixed Android 1.5 support bug
In Android 1.5, setTag() is used by SimpleAdapter, which conflicted with how aguamarina used it, which actually explains a lot about why the exception was a ClassCastException. I solved this conflict by using a hashmap. I also removed the warning shown when running on Android 1.5 and a now unused image of bugdroid.