Moved "deep" variable to BaseManagement
The "deep" variable is used to track the level within the categories to go back when using the back button. I moved it to BaseManagement to keep track of it after an app is installed or uninstalled, and to use it to fix a problem where the carousel appears where it's not supposed to after installing or uninstalling an application.